Add support for linking apps to environments and the reverse.
This adds support for determining whether or not an environment is linked to
its constituent apps, or linking apps to their environments.

The former is necessary to be able to bring up all the apps in an environment,
and the latter is necessary to bring up apps with their dependencies in the
environment.